backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
opencv-python-headless>=4.9.0
numpy>=2.0.0
google-genai>=1.0.0
httpx>=0.27.0
shapely>=2.0.7
build123d>=0.8.0
numpy-stl>=3.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0
